integrations/crdgen: Support loading instructions from a file for easier debugging
